Move CustomLogin action to BackendDirector from SignWidget
