+++ Assets/_Scripts/UI/SignWidget.cs
@@ -2,8 +2,6 @@
 using ChessCrush.Game;
 using ChessCrush.OperationResultCode;
 using System;
-using System.Collections;
-using System.Threading.Tasks;
 using UniRx;
 using UnityEngine;
 using UnityEngine.UI;
@@ -44,6 +42,7 @@ namespace ChessCrush.UI
         private Button exitButton;
 
         private StartSceneDirector startSceneDirector;
+        private BackendDirector backendDirector;
 
         private void Awake()
         {
@@ -57,6 +56,7 @@ namespace ChessCrush.UI
         private void Start()
         {
             startSceneDirector = Director.instance.GetSubDirector<StartSceneDirector>();
+            backendDirector = Director.instance.GetSubDirector<BackendDirector>();
         }
 
         private void OnEnable()
@@ -79,42 +79,14 @@ namespace ChessCrush.UI
                 return;
             }
 
-            var success = new ReactiveProperty<bool>();
-            var bro = new BackendReturnObject();
+            backendDirector.CustomLogin(signInIDInputField.text, signInPWInputField.text, SetAfterSignIn, str => MessageBoxUI.UseWithComponent(str));
+        }
 
-            Backend.BMember.CustomLogin(signInIDInputField.text, signInPWInputField.text, c =>
-              {
-                  bro = c;
-                  success.Value = true;
-              });
-
-            success.ObserveOnMainThread().Subscribe(value =>
-            {
-                if(value)
-                {
-                    var saveToken = Backend.BMember.SaveToken(bro);
-                    if(bro.IsSuccess())
-                    {
-                        MessageBoxUI.UseWithComponent("Success to sign in");
-                        startSceneDirector.signedIn.Value = true;
-                        gameObject.SetActive(false);
-                    }
-                    else
-                    {
-                        switch((SignInCode)Convert.ToInt32(bro.GetStatusCode()))
-                        {
-                            case SignInCode.BadUnauthorizedException:
-                                MessageBoxUI.UseWithComponent("Failed to sign in: wrong id or password");
-                                break;
-                            case SignInCode.Blocked:
-                                MessageBoxUI.UseWithComponent("Failed to sign in: blocked user");
-                                break;
-                            case SignInCode.Etc:
-                                return;
-                        }
-                    }
-                }
-            });
+        private void SetAfterSignIn()
+        {
+            MessageBoxUI.UseWithComponent("Success to sign in");
+            startSceneDirector.signedIn.Value = true;
+            gameObject.SetActive(false);
         }
 
         private void SubscribeSignInSignUpButton()
@@ -138,7 +110,7 @@ namespace ChessCrush.UI
                 return;
             }
 
-            Director.instance.GetSubDirector<BackendDirector>().CustomSignUp(signUpIDInputField.text, signUpPWInputField.text, SetAfterSignUp, str => MessageBoxUI.UseWithComponent(str));
+            backendDirector.CustomSignUp(signUpIDInputField.text, signUpPWInputField.text, SetAfterSignUp, str => MessageBoxUI.UseWithComponent(str));
         }
 
         private void SetAfterSignUp()
